Allow a port number in the registry address
When splitting container tags on ":", the build scripts now take into
account the possibility of 3 items getting returned if the
registry address includes a port number, in which case the first two
components re-joined to form the tag name. The version is always
extracted as the last element of the list instead of counting
from the head.